Improve Terraform inventory options (#55)
Improve Terraform inventory options

SUMMARY
This pull request includes changes to the Terraform inventory plugin.
The three changes are:

Added search_child_modules option
Allow project_path to accept a list of paths.
Set default of state_file to a blank string

All changes in this pull request are backwards compatible with current configurations.
ISSUE TYPE

Feature Pull Request

COMPONENT NAME
cloud.terraform.terraform_provider
ADDITIONAL INFORMATION
search_child_modules option
Terraform modules are often used to consolidate code into portable parts. The inventory plugin currently ignores resources created by modules. This PR adds a boolean option search_child_modules (default: true) to also search through child modules.
I believe this to be a sane default as the ansible_host resource needs to be explicitly added within the module.
The pull request #51 also addresses this, but does not allow it to be disabled. I also dislike how this PR creates TerraformRootModuleResource objects for the child module resources.
In my PR I have split TerraformRootModuleResource into TerraformRootModuleResource and TerraformChildModuleResource, both inheriting a new TerraformModuleResource. This is useful for if future differing behaviours are required. Tests, type hints and imports are adjusted to accept the new typing.
project_path list
This PR also allows project_path to be a list. This is useful for when you are using multiple Terraform projects, but still want to use them in a single Ansible project (for example in a multi-cloud setup, splitting large projects or separation of concerns in organisations)
If an ansible_host resource with the same Ansible inventory name is found in multiple projects it will return a warning and exit to prevent unpredictable situations.
If the given type is a string it will turn it into a single item list, so this is fully backwards compatible.
state_file default blank string
This removes the default value of "terraform.tfstate" of state_file. This is unnecessary as Terraform itself will automatically try to find this file if the statefile is not explicitly given. By defaulting to a blank string it is not given.
This also improves setups with a remote state (Storing on GitLab, S3 bucket, Azure data, etc), as giving the statefile will cause Terraform to not request the state at the backend. This is especially important in combination with project_path as a list where multiple projects can be used with the state file at potentially differing locations, Terraform is able to resolve this itself.
This is backwards compatible as well, as Terraform will try to read terraform.tfstate without it being explicitly given.
I've set the version of version_added in the options documentation to 1.2.0 to avoid confusion on whether the project_path accepts a list.

Reviewed-by: Bikouo Aubin
Reviewed-by: diademiemi <emilia@diademiemi.me>
Reviewed-by: Polona Mihalic <mihalic.polona@gmail.com>
Reviewed-by: Mike Graves <mgraves@redhat.com>