Add GH release automation workflows
This adds two new GHA workflows for automating the creation of GH
releases from tags. It uses the actions from the amazon.aws repo and
adds the same workflows for automatically creating a GH release when a
new tag is created and for manually tiggering the workflow for an
existing tag.