Add 2.14 to supported ansible versions for lint (#156)
SUMMARY

Ansible lint checks that the requires_ansible resolves to a supported version of Ansible. This is a problem for major versions we still support that have an older minimum required version of ansible. Version 24.6.0 of ansible-lint added a new config option to relax the requires_ansible version check.
This PR is only against the stable-2 branch.

ISSUE TYPE


Bugfix Pull Request

COMPONENT NAME

ADDITIONAL INFORMATION

Reviewed-by: Alina Buzachis
Reviewed-by: Bikouo Aubin