AAP46677 - Update workspace logic (#200)
SUMMARY
Refers to AAP-46677
Replaces #194 and #197

Breaking change: Remove the default value default for the workspace argument. When the user runs terraform init, it may use a different workspace value. When running apply on the directory, we should use the current workspace value (retrieved using terraform workspace show) instead of always defaulting to default. In case not set, the CLI will automatically default to the default value.
Update workspace logic

In case the user has provided a workspace argument, this will be set into the TF_WORKSPACE environment for all terraform commands. If the value differs from the one in the cloud configuration block, the CLI will raise an error.
In case the user has not provided a workspace argument the CLI will determine the workspace using either the cloud configuration block or the default value.



ISSUE TYPE


Feature Pull Request

COMPONENT NAME

terraform

Reviewed-by: Abhishekism9450
Reviewed-by: Shashank Venkat
Reviewed-by: Bianca Henderson <beeankha@gmail.com>